split up firmware into multiple files
